Develop ami

+++ todo-app/servers/servers.js
@@ -62,27 +62,22 @@ app.get('/todos/:id', authenticate, (req, res) => {
     });
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-    let id = req.params.id;
-
-    let isValid = ObjectID.isValid(id);
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id;
+    const isValid = ObjectID.isValid(id);
     if(!isValid) {
         return res.status(404).send({error: 'invalid id'});
     }
-
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    }).then((todo) => {
+    try{
+        const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
         if(!todo) {
             return res.status(404).send({error: 'id not found'});
         }
         res.send({todo});
-    }).catch((err) => {
+    }catch(e) {
         res.status(400).send({});
-    });
+    };
 });
-
 
 app.patch('/todos/:id', authenticate, (req, res) =>{
     let id = req.params.id;
@@ -118,43 +113,40 @@ app.patch('/todos/:id', authenticate, (req, res) =>{
 });
 
 
-app.post('/users', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password']);
-    let user = new User(body);
-
-    user.save()
-        .then(() => {
-            return user.generateAuthToken();
-        })
-        .then((token) => {
-            res.header('x-auth', token).send(user);
-        })
-        .catch((err) => {
-            res.status(400).send(err);
-        });
+app.post('/users', async (req, res) => {
+    try{
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = new User(body);
+        await user.save();
+        const token = user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    }catch(e) {
+        res.status(400).send(e);
+    }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-    let body = _.pick(req.body, ['email', 'password']);
-    User.findByCredentials(body.email, body.password).then((user) => {
-        return user.generateAuthToken().then((token) => {
-            res.header('x-auth', token).send(user);
-        })
-    }).catch((err) => {
-        res.status(400).send(err);
-    });
+app.post('/users/login', async (req, res) => {
+    try{
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    }catch(e) {
+        res.status(400).send(e);
+    }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try{
+        await req.user.removeToken(req.token);
         res.status(200).send();
-    }).catch(() => {
+    }catch(e) {
         res.status(400).send();
-    });
+    }
 });
